[Win32][VM][X07] Fix FTBFS.

+++ source/src/vm/x07/io.cpp
@@ -1346,7 +1346,7 @@ bool IO::load_state(FILEIO* state_fio)
 		cmt_fio->Fopen(rec_file_path, FILEIO_READ_WRITE_NEW_BINARY);
 		while(length_tmp != 0) {
 			uint8 buffer_tmp[1024];
-			int length_rw = min((int)length_tmp, sizeof(buffer_tmp));
+			int length_rw = min((int)length_tmp, (int)sizeof(buffer_tmp));
 			state_fio->Fread(buffer_tmp, length_rw, 1);
 			if(cmt_fio->IsOpened()) {
 				cmt_fio->Fwrite(buffer_tmp, length_rw, 1);
